add db identity table
