+++ src/main/java/com/iemr/common/identity/config/PrimaryDBConfig.java
@@ -42,7 +42,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @Configuration
 @EnableTransactionManagement
-@EnableJpaRepositories(entityManagerFactoryRef = "entityManagerFactory", basePackages = {"com.iemr.common.identity.domain",
+@EnableJpaRepositories(entityManagerFactoryRef = "entityManagerFactory", basePackages = {"com.iemr.common.identity.domain.identity",
         "com.iemr.common.identity.repo"})
 public class PrimaryDBConfig {
 
@@ -77,7 +77,7 @@ public class PrimaryDBConfig {
     @Bean(name = "entityManagerFactory")
     public LocalContainerEntityManagerFactoryBean entityManagerFactory(EntityManagerFactoryBuilder builder,
                                                                        @Qualifier("dataSource") DataSource dataSource) {
-        return builder.dataSource(dataSource).packages("com.iemr.common.identity.domain").persistenceUnit("db_identity").build();
+        return builder.dataSource(dataSource).packages("com.iemr.common.identity.domain.identity").persistenceUnit("db_identity").build();
     }
 
     @Primary
