+++ src/main/java/com/iemr/common/identity/domain/identity/MBeneficiaryAccount.java
@@ -19,7 +19,7 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.iemr.common.identity.domain;
+package com.iemr.common.identity.domain.identity;
 
 import java.math.BigInteger;
 import java.sql.Timestamp;
