+++ src/main/java/com/iemr/common/identity/domain/identity/MBeneficiaryImage.java
@@ -19,9 +19,8 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.iemr.common.identity.domain;
+package com.iemr.common.identity.domain.identity;
 
-import java.io.Serializable;
 import java.math.BigInteger;
 import java.sql.Timestamp;
 
@@ -30,76 +29,62 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.NamedQuery;
 import jakarta.persistence.Table;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import com.google.gson.annotations.Expose;
 
 import lombok.Data;
 
-/**
- * The persistent class for the m_beneficiaryfamilymapping database table.
- * 
- */
 @Entity
-@Table(name = "i_beneficiaryfamilymapping")
-@NamedQuery(name = "MBeneficiaryfamilymapping.findAll", query = "SELECT m FROM MBeneficiaryfamilymapping m")
+@Table(name = "i_beneficiaryimage")
 @Data
-// @SelectBeforeUpdate(value=true)
-// @DynamicUpdate(value=true)
-public class MBeneficiaryfamilymapping implements Serializable {
-	private static final long serialVersionUID = 1L;
-
+public class MBeneficiaryImage {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(unique = true, nullable = false)
-	private BigInteger benFamilyMapId;
-
-	private BigInteger associatedBenRegId;
-
-	@Column(nullable = false, updatable = false, insertable = true, length = 50)
-	private String createdBy;
-
-	@Column(name = "CreatedDate", insertable = false, updatable = false)
-	private Timestamp createdDate;
-
+	@Column(name = "BenImageId")
+	private BigInteger benImageId;
+	@Column(name = "BenImage")
+	private String benImage;
+	@Column(name = "BenBiometric")
+	private String benBiometric;
 	@Column(name = "Deleted", insertable = false, updatable = true)
-	private Boolean deleted;
-
-	private Boolean isEmergencyContact;
-
-	@Column(name = "LastModDate", insertable = false, updatable = false)
+	private Boolean deleted = false;
+	@Column(name = "Processed", insertable = false)
+	private String processed;
+	@Column(name = "CreatedBy", updatable = false)
+	private String createdBy;
+	@Column(name = "CreatedDate",updatable = false)
+	@JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
+	private Timestamp createdDate;
+	@Column(name = "Reserved")
+	private Boolean reserved;
+	@Column(name = "ReservedFor")
+	private String reservedFor;
+	@Column(name = "ReservedOn")
+	private String reservedOn;
+	@Column(name = "ReservedById")
+	private Integer reservedById;
+	@Column(name = "ModifiedBy")
+	private String modifiedBy;
+	@Column(name = "LastModDate")
+	@JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
 	private Timestamp lastModDate;
 
-	@Column(length = 50)
-	private String modifiedBy;
-
-	@Column(nullable = false, length = 4)
-	private String processed = "N";
-
-	@Column(length = 15)
-	private String relationshipToSelf;
-
-	@Column
-	private Integer relationshipID;
-
-	private Boolean reserved;
-
-	private Integer reservedById;
-
-	@Column(length = 45)
-	private String reservedFor;
-
-	private Timestamp reservedOn;
-
-
-	@Column(name = "BenMapId", insertable = true, updatable = true)
-	private BigInteger benMapId;
+	@Column(name = "VehicalNo")
+	private String vehicalNo;
+	@Column(name = "SyncedBy")
+	private String syncedBy;
+	@Column(name = "SyncedDate", unique = true, nullable = false)
+	@JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
+	private Timestamp syncedDate;
+	@Column(name = "ReservedForChange")
+	private String reservedForChange;
 
 	// new column added for data sync
 	// 17-06-2018
 	@Expose
-	@Column(name = "VanID", updatable = false)
+	@Column(name = "vanID", updatable = false)
 	private Integer vanID;
 	@Expose
 	@Column(name = "parkingPlaceID", updatable = false)
@@ -107,7 +92,7 @@ public class MBeneficiaryfamilymapping implements Serializable {
 
 	@Expose
 	@Column(name = "VanSerialNo", updatable = false)
-	private BigInteger vanSerialNo;
+	private Long vanSerialNo;
 
 	// END OF new column added for data sync
