+++ src/main/java/com/iemr/common/identity/domain/identity/MBeneficiaryconsent.java
@@ -19,14 +19,11 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.iemr.common.identity.domain;
+package com.iemr.common.identity.domain.identity;
 
 import java.io.Serializable;
 import java.math.BigInteger;
 import java.sql.Timestamp;
-
-
-import com.google.gson.annotations.Expose;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -35,26 +32,28 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.NamedQuery;
 import jakarta.persistence.Table;
-import jakarta.validation.constraints.Email;
+
+import com.google.gson.annotations.Expose;
+
 import lombok.Data;
 
 /**
- * The persistent class for the m_beneficiarycontacts database table.
+ * The persistent class for the m_beneficiaryconsent database table.
  * 
  */
 @Entity
-@Table(name = "i_beneficiarycontacts")
-@NamedQuery(name = "MBeneficiarycontact.findAll", query = "SELECT m FROM MBeneficiarycontact m")
+@Table(name = "i_beneficiaryconsent")
+@NamedQuery(name = "MBeneficiaryconsent.findAll", query = "SELECT m FROM MBeneficiaryconsent m")
 @Data
-public class MBeneficiarycontact implements Serializable {
+public class MBeneficiaryconsent implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(unique = true, nullable = false)
-	private BigInteger benContactsID;
+	private BigInteger benConsentID;
 
-	@Column(nullable = false, length = 50, updatable = false)
+	@Column(nullable = false, length = 50)
 	private String createdBy;
 
 	@Column(name = "CreatedDate",updatable = false)
@@ -62,62 +61,10 @@ public class MBeneficiarycontact implements Serializable {
 
 	private Boolean deleted = false;
 
-	@Email
-	@Column(length = 75)
-	private String emailId;
-
-	@Column(length = 25)
-	private String emergencyContactNum;
-
-	@Column(length = 15)
-	private String emergencyContactTyp;
-
 	private Timestamp lastModDate;
 
 	@Column(length = 50)
 	private String modifiedBy;
-
-	@Column(length = 25)
-	private String phoneNum1;
-
-	@Column(length = 25)
-	private String phoneNum2;
-
-	@Column(length = 25)
-	private String phoneNum3;
-
-	@Column(length = 25)
-	private String phoneNum4;
-
-	@Column(length = 25)
-	private String phoneNum5;
-
-	@Column(length = 15)
-	private String phoneTyp1;
-
-	@Column(length = 15)
-	private String phoneTyp2;
-
-	@Column(length = 15)
-	private String phoneTyp3;
-
-	@Column(length = 15)
-	private String phoneTyp4;
-
-	@Column(length = 15)
-	private String phoneTyp5;
-
-	@Column(nullable = false, length = 25)
-	private String preferredPhoneNum;
-
-	@Column(length = 15)
-	private String preferredPhoneTyp;
-
-	@Column(nullable = false, length = 25)
-	private String preferredSMSPhoneNum;
-
-	@Column(length = 15)
-	private String preferredSMSPhoneTyp;
 
 	@Column(nullable = false, length = 4)
 	private String processed = "N";
@@ -130,6 +77,40 @@ public class MBeneficiarycontact implements Serializable {
 	private String reservedFor;
 
 	private Timestamp reservedOn;
+
+	private Boolean shareAnonymousWithGovt;
+
+	private Boolean shareAnonymousWithMedicalCommunity;
+
+	private Boolean shareAnonymousWithNGO;
+
+	private Boolean shareMedicalDetailsForMedicalStudy;
+
+	private Boolean shareMedicalDetailsWithDoctor;
+
+	private Boolean shareMedicalDetailsWithFamily;
+
+	private Boolean shareMedicalDetailsWithFriends;
+
+	private Boolean shareMedicalDetailsWithGovt;
+
+	private Boolean shareMedicalDetailsWithNGO;
+
+	private Boolean shareMedicalDetailsWithSpouse;
+
+	private Boolean sharePersonalDetailsForMedicalStudy;
+
+	private Boolean sharePersonalDetailsWithDoctor;
+
+	private Boolean sharePersonalDetailsWithFamily;
+
+	private Boolean sharePersonalDetailsWithFriends;
+
+	private Boolean sharePersonalDetailsWithGovt;
+
+	private Boolean sharePersonalDetailsWithNGO;
+
+	private Boolean sharePersonalDetailsWithSpouse;
 
 	// new column added for data sync
 	// 17-06-2018
@@ -145,4 +126,9 @@ public class MBeneficiarycontact implements Serializable {
 	private BigInteger vanSerialNo;
 
 	// END OF new column added for data sync
+
+	@Expose
+	@Column(name = "BenConsent", updatable = false)
+	private Boolean beneficiaryConsent;
+
 }
