+++ src/main/java/com/iemr/common/identity/domain/identity/MBeneficiarycontact.java
@@ -19,11 +19,14 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.iemr.common.identity.domain;
+package com.iemr.common.identity.domain.identity;
 
 import java.io.Serializable;
 import java.math.BigInteger;
 import java.sql.Timestamp;
+
+
+import com.google.gson.annotations.Expose;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -32,28 +35,26 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.NamedQuery;
 import jakarta.persistence.Table;
-
-import com.google.gson.annotations.Expose;
-
+import jakarta.validation.constraints.Email;
 import lombok.Data;
 
 /**
- * The persistent class for the m_beneficiaryconsent database table.
+ * The persistent class for the m_beneficiarycontacts database table.
  * 
  */
 @Entity
-@Table(name = "i_beneficiaryconsent")
-@NamedQuery(name = "MBeneficiaryconsent.findAll", query = "SELECT m FROM MBeneficiaryconsent m")
+@Table(name = "i_beneficiarycontacts")
+@NamedQuery(name = "MBeneficiarycontact.findAll", query = "SELECT m FROM MBeneficiarycontact m")
 @Data
-public class MBeneficiaryconsent implements Serializable {
+public class MBeneficiarycontact implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(unique = true, nullable = false)
-	private BigInteger benConsentID;
+	private BigInteger benContactsID;
 
-	@Column(nullable = false, length = 50)
+	@Column(nullable = false, length = 50, updatable = false)
 	private String createdBy;
 
 	@Column(name = "CreatedDate",updatable = false)
@@ -61,10 +62,62 @@ public class MBeneficiaryconsent implements Serializable {
 
 	private Boolean deleted = false;
 
+	@Email
+	@Column(length = 75)
+	private String emailId;
+
+	@Column(length = 25)
+	private String emergencyContactNum;
+
+	@Column(length = 15)
+	private String emergencyContactTyp;
+
 	private Timestamp lastModDate;
 
 	@Column(length = 50)
 	private String modifiedBy;
+
+	@Column(length = 25)
+	private String phoneNum1;
+
+	@Column(length = 25)
+	private String phoneNum2;
+
+	@Column(length = 25)
+	private String phoneNum3;
+
+	@Column(length = 25)
+	private String phoneNum4;
+
+	@Column(length = 25)
+	private String phoneNum5;
+
+	@Column(length = 15)
+	private String phoneTyp1;
+
+	@Column(length = 15)
+	private String phoneTyp2;
+
+	@Column(length = 15)
+	private String phoneTyp3;
+
+	@Column(length = 15)
+	private String phoneTyp4;
+
+	@Column(length = 15)
+	private String phoneTyp5;
+
+	@Column(nullable = false, length = 25)
+	private String preferredPhoneNum;
+
+	@Column(length = 15)
+	private String preferredPhoneTyp;
+
+	@Column(nullable = false, length = 25)
+	private String preferredSMSPhoneNum;
+
+	@Column(length = 15)
+	private String preferredSMSPhoneTyp;
 
 	@Column(nullable = false, length = 4)
 	private String processed = "N";
@@ -77,40 +130,6 @@ public class MBeneficiaryconsent implements Serializable {
 	private String reservedFor;
 
 	private Timestamp reservedOn;
-
-	private Boolean shareAnonymousWithGovt;
-
-	private Boolean shareAnonymousWithMedicalCommunity;
-
-	private Boolean shareAnonymousWithNGO;
-
-	private Boolean shareMedicalDetailsForMedicalStudy;
-
-	private Boolean shareMedicalDetailsWithDoctor;
-
-	private Boolean shareMedicalDetailsWithFamily;
-
-	private Boolean shareMedicalDetailsWithFriends;
-
-	private Boolean shareMedicalDetailsWithGovt;
-
-	private Boolean shareMedicalDetailsWithNGO;
-
-	private Boolean shareMedicalDetailsWithSpouse;
-
-	private Boolean sharePersonalDetailsForMedicalStudy;
-
-	private Boolean sharePersonalDetailsWithDoctor;
-
-	private Boolean sharePersonalDetailsWithFamily;
-
-	private Boolean sharePersonalDetailsWithFriends;
-
-	private Boolean sharePersonalDetailsWithGovt;
-
-	private Boolean sharePersonalDetailsWithNGO;
-
-	private Boolean sharePersonalDetailsWithSpouse;
 
 	// new column added for data sync
 	// 17-06-2018
@@ -126,9 +145,4 @@ public class MBeneficiaryconsent implements Serializable {
 	private BigInteger vanSerialNo;
 
 	// END OF new column added for data sync
-
-	@Expose
-	@Column(name = "BenConsent", updatable = false)
-	private Boolean beneficiaryConsent;
-
 }
