+++ src/main/java/com/iemr/common/identity/domain/identity/MBeneficiaryregidmapping.java
@@ -19,8 +19,9 @@
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see https://www.gnu.org/licenses/.
 */
-package com.iemr.common.identity.domain;
+package com.iemr.common.identity.domain.identity;
 
+import java.io.Serializable;
 import java.math.BigInteger;
 import java.sql.Timestamp;
 
@@ -29,71 +30,72 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.NamedQuery;
 import jakarta.persistence.Table;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import com.google.gson.annotations.Expose;
 
 import lombok.Data;
 
+/**
+ * The persistent class for the m_beneficiaryregid database table.
+ * 
+ */
 @Entity
-@Table(name = "i_beneficiaryimage")
+@Table(name = "m_beneficiaryregidmapping")
+@NamedQuery(name = "MBeneficiaryregidmapping.findAll", query = "SELECT m FROM MBeneficiaryregidmapping m")
 @Data
-public class MBeneficiaryImage {
+public class MBeneficiaryregidmapping implements Serializable {
+	private static final long serialVersionUID = 1L;
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "BenImageId")
-	private BigInteger benImageId;
-	@Column(name = "BenImage")
-	private String benImage;
-	@Column(name = "BenBiometric")
-	private String benBiometric;
-	@Column(name = "Deleted", insertable = false, updatable = true)
-	private Boolean deleted = false;
-	@Column(name = "Processed", insertable = false)
-	private String processed;
-	@Column(name = "CreatedBy", updatable = false)
+	@Column(unique = true, nullable = false)
+	private BigInteger benRegId;
+
+	@Column(nullable = false)
+	private BigInteger beneficiaryID;
+
+	@Column(length = 30)
 	private String createdBy;
+
 	@Column(name = "CreatedDate",updatable = false)
-	@JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
 	private Timestamp createdDate;
-	@Column(name = "Reserved")
+
+	private Boolean deleted = false;
+
+	private Boolean provisioned;
+
+	private Integer providerServiceMapID;
+
 	private Boolean reserved;
-	@Column(name = "ReservedFor")
-	private String reservedFor;
-	@Column(name = "ReservedOn")
-	private String reservedOn;
-	@Column(name = "ReservedById")
+
 	private Integer reservedById;
-	@Column(name = "ModifiedBy")
-	private String modifiedBy;
-	@Column(name = "LastModDate")
-	@JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-	private Timestamp lastModDate;
+
+	@Column(length = 45)
+	private String reservedFor;
+
+	private Timestamp reservedOn;
 
 	@Column(name = "VehicalNo")
 	private String vehicalNo;
-	@Column(name = "SyncedBy")
+
+
 	private String syncedBy;
-	@Column(name = "SyncedDate", unique = true, nullable = false)
-	@JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
+
 	private Timestamp syncedDate;
-	@Column(name = "ReservedForChange")
+
 	private String reservedForChange;
 
-	// new column added for data sync
-	// 17-06-2018
 	@Expose
-	@Column(name = "vanID", updatable = false)
+	@Column(name = "vanID")
 	private Integer vanID;
 	@Expose
-	@Column(name = "parkingPlaceID", updatable = false)
+	@Column(name = "parkingPlaceID")
 	private Integer parkingPlaceID;
 
 	@Expose
-	@Column(name = "VanSerialNo", updatable = false)
-	private Long vanSerialNo;
-
-	// END OF new column added for data sync
+	@Column(name = "VanSerialNo")
+	private BigInteger vanSerialNo;
 
 }
