+++ src/main/java/com/iemr/common/identity/dto/IdentityEditDTO.java
@@ -25,9 +25,9 @@ import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.List;
 
-import com.iemr.common.identity.domain.Address;
-import com.iemr.common.identity.domain.Contact;
-import com.iemr.common.identity.domain.Identity;
+import com.iemr.common.identity.domain.identity.Address;
+import com.iemr.common.identity.domain.identity.Contact;
+import com.iemr.common.identity.domain.identity.Identity;
 
 import lombok.Data;
