+++ src/main/java/com/iemr/common/identity/mapper/IdentityEditMapper.java
@@ -30,17 +30,17 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
-import com.iemr.common.identity.domain.Identity;
-import com.iemr.common.identity.domain.MBeneficiaryAccount;
-import com.iemr.common.identity.domain.MBeneficiaryImage;
-import com.iemr.common.identity.domain.MBeneficiaryaddress;
-import com.iemr.common.identity.domain.MBeneficiaryconsent;
-import com.iemr.common.identity.domain.MBeneficiarycontact;
-import com.iemr.common.identity.domain.MBeneficiarydetail;
-import com.iemr.common.identity.domain.MBeneficiaryfamilymapping;
-import com.iemr.common.identity.domain.MBeneficiaryidentity;
-import com.iemr.common.identity.domain.MBeneficiarymapping;
-import com.iemr.common.identity.domain.MBeneficiaryservicemapping;
+import com.iemr.common.identity.domain.identity.Identity;
+import com.iemr.common.identity.domain.identity.MBeneficiaryAccount;
+import com.iemr.common.identity.domain.identity.MBeneficiaryImage;
+import com.iemr.common.identity.domain.identity.MBeneficiaryaddress;
+import com.iemr.common.identity.domain.identity.MBeneficiaryconsent;
+import com.iemr.common.identity.domain.identity.MBeneficiarycontact;
+import com.iemr.common.identity.domain.identity.MBeneficiarydetail;
+import com.iemr.common.identity.domain.identity.MBeneficiaryfamilymapping;
+import com.iemr.common.identity.domain.identity.MBeneficiaryidentity;
+import com.iemr.common.identity.domain.identity.MBeneficiarymapping;
+import com.iemr.common.identity.domain.identity.MBeneficiaryservicemapping;
 import com.iemr.common.identity.dto.BenDetailDTO;
 import com.iemr.common.identity.dto.BenFamilyDTO;
 import com.iemr.common.identity.dto.BenIdentityDTO;
