+++ src/main/java/com/iemr/common/identity/repo/BenDetailRepo.java
@@ -32,7 +32,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.iemr.common.identity.domain.MBeneficiarydetail;
+import com.iemr.common.identity.domain.identity.MBeneficiarydetail;
 
 import jakarta.transaction.Transactional;
