+++ src/main/java/com/iemr/common/identity/repo/BenRegIdMappingRepo.java
@@ -31,7 +31,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import com.iemr.common.identity.domain.MBeneficiaryregidmapping;
+import com.iemr.common.identity.domain.identity.MBeneficiaryregidmapping;
 
 import jakarta.transaction.Transactional;
