+++ src/main/java/com/iemr/common/identity/service/IdentityService.java
@@ -50,19 +50,19 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.google.gson.LongSerializationPolicy;
 import com.iemr.common.identity.data.rmnch.RMNCHBeneficiaryDetailsRmnch;
-import com.iemr.common.identity.domain.Address;
-import com.iemr.common.identity.domain.MBeneficiaryAccount;
-import com.iemr.common.identity.domain.MBeneficiaryImage;
-import com.iemr.common.identity.domain.MBeneficiaryaddress;
-import com.iemr.common.identity.domain.MBeneficiaryconsent;
-import com.iemr.common.identity.domain.MBeneficiarycontact;
-import com.iemr.common.identity.domain.MBeneficiarydetail;
-import com.iemr.common.identity.domain.MBeneficiaryfamilymapping;
-import com.iemr.common.identity.domain.MBeneficiaryidentity;
-import com.iemr.common.identity.domain.MBeneficiarymapping;
-import com.iemr.common.identity.domain.MBeneficiaryregidmapping;
-import com.iemr.common.identity.domain.MBeneficiaryservicemapping;
-import com.iemr.common.identity.domain.VBenAdvanceSearch;
+import com.iemr.common.identity.domain.identity.Address;
+import com.iemr.common.identity.domain.identity.MBeneficiaryAccount;
+import com.iemr.common.identity.domain.identity.MBeneficiaryImage;
+import com.iemr.common.identity.domain.identity.MBeneficiaryaddress;
+import com.iemr.common.identity.domain.identity.MBeneficiaryconsent;
+import com.iemr.common.identity.domain.identity.MBeneficiarycontact;
+import com.iemr.common.identity.domain.identity.MBeneficiarydetail;
+import com.iemr.common.identity.domain.identity.MBeneficiaryfamilymapping;
+import com.iemr.common.identity.domain.identity.MBeneficiaryidentity;
+import com.iemr.common.identity.domain.identity.MBeneficiarymapping;
+import com.iemr.common.identity.domain.identity.MBeneficiaryregidmapping;
+import com.iemr.common.identity.domain.identity.MBeneficiaryservicemapping;
+import com.iemr.common.identity.domain.identity.VBenAdvanceSearch;
 import com.iemr.common.identity.dto.AbhaAddressDTO;
 import com.iemr.common.identity.dto.BenIdImportDTO;
 import com.iemr.common.identity.dto.BeneficiariesDTO;
@@ -95,7 +95,6 @@ import com.iemr.common.identity.repo.rmnch.RMNCHBeneficiaryDetailsRmnchRepo;
 import com.iemr.common.identity.service.elasticsearch.ElasticsearchService;
 import com.iemr.common.identity.utils.mapper.OutputMapper;
 import com.iemr.common.identity.utils.response.OutputResponse;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.beans.factory.annotation.Value;
 
 import com.iemr.common.identity.service.elasticsearch.BeneficiaryElasticsearchIndexUpdater;
