+++ src/main/java/com/iemr/common/identity/service/familyTagging/FamilyTagServiceImpl.java
@@ -37,8 +37,8 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.iemr.common.identity.data.familyTagging.BenFamilyMapping;
 import com.iemr.common.identity.data.familyTagging.FamilyMembers;
-import com.iemr.common.identity.domain.MBeneficiarydetail;
-import com.iemr.common.identity.domain.MBeneficiarymapping;
+import com.iemr.common.identity.domain.identity.MBeneficiarydetail;
+import com.iemr.common.identity.domain.identity.MBeneficiarymapping;
 import com.iemr.common.identity.exception.IEMRException;
 import com.iemr.common.identity.repo.BenDetailRepo;
 import com.iemr.common.identity.repo.BenMappingRepo;
